-frounding-math is used everywhere when using CGAL
Added GetClusterIDs() method in libDistributedClustering API

Added Size() And Flush() methods to HullModel

Changed ours CGAL's kernel precision from float to double

Moved the generation of output files from the generic ClusteringBackend to its
derived classes (new virtual method ProcessResults())

Tree_DBSCAN_env.sh now updates the PATH variable

Added script 'draw_hulls'

Added some missing BSC headers and debug messages

+++ src/libClustering/CGAL_Kernel/MyConstruct_coord_iterator.h
@@ -3,14 +3,14 @@
 
 class MyConstruct_coord_iterator {
 public:
-  const float* operator()(const MyPointC2& p)
+  const double* operator()(const MyPointC2& p)
   {
     return &p.x();
   }
 
-  const float* operator()(const MyPointC2& p, int)
+  const double* operator()(const MyPointC2& p, int)
   {
-    const float* pyptr = &p.y();
+    const double* pyptr = &p.y();
     pyptr++;
     return pyptr;
   }
